Restructure PowerDNS DSP based on layout for CSCGlobal
